update README.md, enforce style
this updates the README.md to be a little bit more representative
of what's actually happening (and look a bit nicer).

second i applied code style to the actual code that's written right
now (sysmodule). the closest to code that Gren actually wrote (and
wasn't generated/copied from elsewhere) seemed to be WebKit so I
chose that as the style, but we can obviously change super easy